// src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
// ...
      messages: {
        Row: {
          content: string;
          created_at: string | null;
          id: string;
          read: boolean | null;
          recipient_id: string | null;
          sender_id: string | null;
        };
        Insert: {
          content: string;
          created_at?: string | null;
          id?: string;
          read?: boolean | null;
          recipient_id?: string | null;
          sender_id?: string | null;
        };
        Update: {
          content?: string;
          created_at?: string | null;
          id?: string;
          read?: boolean | null;
          recipient_id?: string | null;
          sender_id?: string | null;
        };
      };
// ...
    };
    Enums: {
      role_enum: "agent" | "client" | "admin";
      tax_status_enum: "completed" | "in-progress" | "pending";
    };
  };
};